apps/transformers/spark-batch/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "spark_batch" ]

[tool.coverage.report]
exclude_lines = [ "if TYPE_CHECKING:" ]
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-fail-under=90 --cov-report html:'../../../coverage/apps/transformers/spark-batch/html' --cov-report xml:'../../../coverage/apps/transformers/spark-batch/coverage.xml' --html='../../../reports/apps/transformers/spark-batch/unittests/html/index.html' --junitxml='../../../reports/apps/transformers/spark-batch/unittests/junit.xml'"

[tool.poetry]
name = "spark-batch"
version = "1.0.0"
description = "Automatically generated by Nx."
authors = [ ]
license = "Proprietary"
readme = "README.md"

  [[tool.poetry.packages]]
  include = "spark_batch"

  [tool.poetry.dependencies]
  python = ">=3.9,<3.11"
  warlock = "^2.0.1"
  pandas = "^2.1.1"
  pyarrow = "^13.0.0"
  grpcio = "^1.58.0"
  protobuf = "^4.24.3"
  grpcio-status = "^1.58.0"
  pyspark = "3.5.0"

    [tool.poetry.dependencies.pylog]
    path = "../../../libs/python/pylog"
    develop = true

    [tool.poetry.dependencies.pyrabbitmq]
    path = "../../../libs/python/pyrabbitmq"
    develop = true

    [tool.poetry.dependencies.pysd]
    path = "../../../libs/python/pysd"
    develop = true

    [tool.poetry.dependencies.config-loader]
    path = "../../../libs/service-setup/python/config-loader"
    develop = true

    [tool.poetry.dependencies.py-repository]
    path = "../../../libs/api-client/python/py-repository"
    develop = true

    [tool.poetry.dependencies.dto-serializer]
    path = "../../../libs/dtos/python/serializer"
    develop = true

    [tool.poetry.dependencies.py-minio]
    path = "../../../libs/storage-client/python/py-minio"
    develop = true

    [tool.poetry.dependencies.dto-gateway]
    path = "../../../libs/dtos/python/dto-gateway"
    develop = true

    [tool.poetry.dependencies.dto-events]
    path = "../../../libs/dtos/python/dto-events"
    develop = true

    [tool.poetry.dependencies.py-lake-catalog]
    path = "../../../libs/api-client/python/py-lake-catalog"
    develop = true

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"
